terminal register and bug fixes;

+++ domain/src/main/java/com/drawint/domain/bo/ActionRegisterBO.java
@@ -9,7 +9,7 @@ public class ActionRegisterBO {
 
     private String code;
 
-    private Short concurrencyLevel;
+    private Integer concurrencyLevel;
 
     private Integer interval;
